+++ config/initializers/contour.rb
@@ -19,10 +19,7 @@ Contour.setup do |config|
   # Enter the items you wish to see in the menu
   config.menu_items = [
     {
-      name: 'Login', display: 'not_signed_in', path: 'new_user_session_path', position: 'right',
-      links: [{ name: 'Sign Up', path: 'new_user_registration_path' },
-              { divider: true },
-              { authentications: true }]
+      name: 'Sign Up', display: 'not_signed_in', path: 'new_user_registration_path', position: 'right'
     },
     {
       name: 'image_tag(current_user.avatar_url(18, "blank"))+" "+current_user.name', eval: true, display: 'signed_in', path: 'settings_path', position: 'right',
@@ -47,10 +44,6 @@ Contour.setup do |config|
     {
       name: 'Users', display: 'signed_in', name: 'Users', path: 'users_path', position: 'left', condition: 'current_user.system_admin?',
       links: [{ name: 'Overall Graph', path: 'overall_graph_users_path', condition: 'current_user.system_admin?' }]
-    },
-    {
-      name: 'About', display: 'always', path: 'about_path', position: 'left',
-      links: []
     }
   ]